requirement.txt
streamlit==1.22.0
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0
numpy==1.24.2